# doc/api/swagger.yaml
basePath: /api/v2
definitions:
...
  database.SimulationModel:
    properties:
      id:
        type: integer
      inputLength:
        description: Number of input signals
        type: integer
      name:
        description: Name of simulation model
        type: string
      outputLength:
        description: Number of output signals
        type: integer
      scenarioID:
        description: ID of Scenario to which simulation model belongs
        type: integer
      simulatorID:
        description: ID of simulator associated with simulation model
        type: integer
      startParameters:
        description: Start parameters of simulation model as JSON
        type: string
    type: object
  database.Simulator:
    properties:
      host:
        description: Host if the simulator
        type: string
      id:
        type: integer
      modelType:
        description: Model type supported by the simulator
        type: string
      properties:
        description: Properties of simulator as JSON string
        type: string
      rawProperties:
        description: Raw properties of simulator as JSON string
        type: string
      state:
        description: State of the simulator
        type: string
      stateUpdateAt:
        description: Time of last state update
        type: string
      uptime:
        description: Uptime of the simulator
        type: integer
      uuid:
        description: UUID of the simulator
        type: string
    type: object
...
  docs.ResponseError:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  docs.ResponseSimulationModels:
    properties:
      models:
        items:
          $ref: '#/definitions/database.SimulationModel'
        type: array
    type: object
  docs.ResponseSimulator:
    properties:
      simulator:
        $ref: '#/definitions/database.Simulator'
        type: object
    type: object
  docs.ResponseSimulators:
    properties:
      simulators:
        items:
          $ref: '#/definitions/database.Simulator'
        type: array
    type: object
...
  simulator.addSimulatorRequest:
    properties:
      simulator:
        $ref: '#/definitions/simulator.validNewSimulator'
        type: object
    type: object
  simulator.updateSimulatorRequest:
    properties:
      simulator:
        $ref: '#/definitions/simulator.validUpdatedSimulator'
        type: object
    type: object
  simulator.validNewSimulator:
    properties:
      Host:
        type: string
      Modeltype:
        type: string
      Properties:
        type: string
      State:
        type: string
      UUID:
        type: string
    required:
    - Host
    - Modeltype
    - Properties
    - UUID
    type: object
  simulator.validUpdatedSimulator:
    properties:
      Host:
        type: string
      Modeltype:
        type: string
      Properties:
        type: string
      State:
        type: string
      UUID:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: sonja.happ@eonerc.rwth-aachen.de
    name: Sonja Happ
  description: |-
    This is the VILLASweb Backend API v2.0.
    Parts of this API are still in development. Please check https://git.rwth-aachen.de/acs/public/villas/web-backend-go for more information.
  license:
    name: GNU GPL 3.0
    url: http://www.gnu.de/documents/gpl-3.0.en.html
  title: VILLASweb Backend API
  version: "2.0"
paths:
...
  /simulators:
    get:
      operationId: getSimulators
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Simulators requested
          schema:
            $ref: '#/definitions/docs.ResponseSimulators'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get all simulators
      tags:
      - simulators
    post:
      consumes:
      - application/json
      operationId: addSimulator
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Simulator to be added
        in: body
        name: inputSimulator
        required: true
        schema:
          $ref: '#/definitions/simulator.addSimulatorRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Simulator that was added
          schema:
            $ref: '#/definitions/docs.ResponseSimulator'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Add a simulator
      tags:
      - simulators
  /simulators/{simulatorID}:
    delete:
      operationId: deleteSimulator
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Simulator ID
        in: path
        name: simulatorID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Simulator that was deleted
          schema:
            $ref: '#/definitions/docs.ResponseSimulator'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Delete a simulator
      tags:
      - simulators
    get:
      operationId: getSimulator
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Simulator ID
        in: path
        name: simulatorID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Simulator that was requested
          schema:
            $ref: '#/definitions/docs.ResponseSimulator'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get simulator
      tags:
      - simulators
    put:
      consumes:
      - application/json
      operationId: updateSimulator
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Simulator to be updated
        in: body
        name: inputSimulator
        required: true
        schema:
          $ref: '#/definitions/simulator.updateSimulatorRequest'
          type: object
      - description: Simulator ID
        in: path
        name: simulatorID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Simulator that was updated
          schema:
            $ref: '#/definitions/docs.ResponseSimulator'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Update a simulator
      tags:
      - simulators
...
  /simulators/{simulatorID}/models:
    get:
      operationId: getModelsOfSimulator
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Simulator ID
        in: path
        name: simulatorID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Simulation models requested by user
          schema:
            $ref: '#/definitions/docs.ResponseSimulationModels'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get all simulation models in which the simulator is used
      tags:
      - simulators
...
swagger: "2.0"